Pass deleteAction as a prop
